Added: popup close by click
