+++ src/components/modal.js
@@ -1,32 +1,46 @@
 import {
-  editProfilePopup,
-  addCardPopup,
-  
-  viewPhotoPopup,
   togglePopup
 } from '../components/utils.js';
 
 import { renderCard } from '../components/card.js';
+
+const popups = document.querySelectorAll('.popup');
 
 const profile = document.querySelector('.profile');
 const profileName = profile.querySelector('.profile__name');
 const profileAbout = profile.querySelector('.profile__about');
 
 const editProfileButton = profile.querySelector('.profile__edit-button');
-const editProfileClose = editProfilePopup.querySelector('.popup__close-button');
 
+const editProfilePopup = document.querySelector('.popup_type_edit-profile');
 const editProfileForm = editProfilePopup.querySelector('.form_type_edit-profile');
 const profileNameInput = editProfileForm.querySelector('.form__item_type_profile-name');
 const profileAboutInput = editProfileForm.querySelector('.form__item_type_profile-about');
 
 const addCardButton = profile.querySelector('.profile__add-button');
-const addCardClose = addCardPopup.querySelector('.popup__close-button');
 
+const addCardPopup = document.querySelector('.popup_type_add-card');
 const addCardForm = addCardPopup.querySelector('.form_type_add-card');
 const cardNameInput = addCardForm.querySelector('.form__item_type_card-name');
 const cardLinkInput = addCardForm.querySelector('.form__item_type_card-link');
 
-const viewPhotoClose = viewPhotoPopup.querySelector('.popup__close-button');
+
+
+// Close popup by click
+
+function closePopupByClick(evt) {
+  if (evt.target === evt.currentTarget || evt.target.classList.contains('popup__close-button')) {
+    togglePopup(evt.currentTarget);
+  }
+}
+
+popups.forEach(function(popup) {
+  popup.addEventListener('click', function (evt) {
+    closePopupByClick(evt);
+  });
+});
+
+// Profile edit
 
 export function enableEditProfile() {
   editProfileButton.addEventListener('click', function () {
@@ -45,11 +59,7 @@ function editProfile (evt) {
 
 editProfileForm.addEventListener('submit', editProfile);
 
-editProfileClose.addEventListener('click', function () {
-  togglePopup(editProfilePopup);
-});
-
-
+// Card add
 
 export function enableAddCard() {
   addCardButton.addEventListener('click', function () {
@@ -65,15 +75,3 @@ export function addCard (evt) {
 }
 
 addCardForm.addEventListener('submit', addCard);
-
-addCardClose.addEventListener('click', function () {
-  togglePopup(addCardPopup);
-});
-
-
-
-
-
-viewPhotoClose.addEventListener('click', function () {
-  togglePopup(viewPhotoPopup);
-});
